Remove obsolete `static_cast` around `std::sin`
Since C++11 there is a `float` overload, so no need to cast from `double`.

+++ src/Trig.cpp
@@ -56,9 +56,7 @@ float NAS2D::angleFromPoints(float x, float y, float x2, float y2)
  */
 Vector<float> NAS2D::angleToDirection(float angle)
 {
-	// static_cast<float> used to suppress warning at possible loss of data. Intentionally
-	// surpressed as we don't need that level of precision.
-	return {static_cast<float>(std::sin(NAS2D::degToRad(angle))), static_cast<float>(-std::cos(NAS2D::degToRad(angle)))};
+	return {std::sin(NAS2D::degToRad(angle)), -std::cos(NAS2D::degToRad(angle))};
 }
 
 
@@ -67,9 +65,7 @@ Vector<float> NAS2D::angleToDirection(float angle)
  */
 Point_2df NAS2D::getDirectionVector(float angle)
 {
-	// static_cast<float> used to suppress warning at possible loss of data. Intentionally
-	// surpressed as we don't need that level of precision.
-	return Point_2df(static_cast<float>(std::sin(NAS2D::degToRad(angle))), static_cast<float>(-std::cos(NAS2D::degToRad(angle))));
+	return Point_2df(std::sin(NAS2D::degToRad(angle)), -std::cos(NAS2D::degToRad(angle)));
 }
